split KernelManager into KernelManager + KernelClient
